move: enum, constant class 디렉토리 이동
