+++ src/main/java/step3/constant/enums/InputStatus.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.constant.enums;
 
 public enum InputStatus {
     MONEY, LOTTO
