+++ src/main/java/step3/domain/LottoManager.java
@@ -1,17 +1,17 @@
 package step3.domain;
 
 import static java.util.Collections.shuffle;
-import static step3.LottoConstant.LOTTO_ELEMENTS_SIZE;
-import static step3.LottoConstant.LOTTO_PRICE;
-import static step3.LottoConstant.LOTTO_VALID_ELEMENTS;
-import static step3.constant.LottoReward.numberToLottoReward;
+import static step3.constant.LottoConstant.LOTTO_ELEMENTS_SIZE;
+import static step3.constant.LottoConstant.LOTTO_PRICE;
+import static step3.constant.LottoConstant.LOTTO_VALID_ELEMENTS;
+import static step3.constant.enums.LottoReward.numberToLottoReward;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.stream.Collectors;
-import step3.constant.LottoReward;
+import step3.constant.enums.LottoReward;
 
 public class LottoManager {
